# Conflicts:
#	frontend/client/public/index.html
#	frontend/client/src/App.css
#	frontend/client/src/App.js
#	frontend/client/src/assets/styles/PortfoliosStyle.css
#	frontend/client/src/assets/styles/SecuritiesStyle.css
#	frontend/client/src/components/cards/PortfolioCard.js
#	frontend/client/src/components/cards/PurchaseInfoCard.js
#	frontend/client/src/components/cards/SecurityCard.js
#	frontend/client/src/components/charts/SecurityPriceChart.js
#	frontend/client/src/components/headers/HeaderMiddle.js
#	frontend/client/src/constants/HeaderLinks.js
#	frontend/client/src/pages/Exchange.js
#	frontend/client/src/pages/ExchangeSecurityInfo.js
#	frontend/client/src/pages/NotFound.js
#	frontend/client/src/pages/Portfolio.js
#	frontend/client/src/pages/PortfolioSecurityInfo.js
#	frontend/client/src/pages/Security.js
#	frontend/client/src/pages/Statistics.js
#	frontend/client/src/util/TypeConverter.js

+++ src/main/java/ru/tink/practice/service/PortfolioService.java
@@ -1,12 +1,18 @@
 package ru.tink.practice.service;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
 import ru.tink.practice.entity.Portfolio;
+import ru.tink.practice.entity.User;
+import ru.tink.practice.exception.PortfolioIsNotEmptyException;
 import ru.tink.practice.exception.PortfolioNotFoundException;
 import ru.tink.practice.repository.PortfolioRepository;
+import ru.tink.practice.security.SecurityUser;
 
+import javax.transaction.Transactional;
 import javax.validation.constraints.NotBlank;
 import java.math.BigDecimal;
 import java.util.List;
@@ -17,13 +23,17 @@ import java.util.List;
 public class PortfolioService {
 
     private final PortfolioRepository portfolioRepository;
+    private final UserService userService;
 
-    public Portfolio savePortfolio(@NotBlank String portfolioName) {
-        return portfolioRepository.save(new Portfolio(portfolioName, BigDecimal.valueOf(0), BigDecimal.valueOf(0)));
+    public Portfolio savePortfolio(SecurityUser securityUser, @NotBlank String portfolioName) {
+        User user = userService.findUserByEmail(securityUser.getEmail());
+        return portfolioRepository.save(new Portfolio(portfolioName,
+                BigDecimal.valueOf(0),
+                user));
     }
 
-    public Portfolio updatePortfolio(Long id, @NotBlank String portfolioName) {
-        Portfolio portfolio = portfolioRepository.findById(id)
+    public Portfolio updatePortfolio(SecurityUser securityUser, Long id, @NotBlank String portfolioName) {
+        Portfolio portfolio = portfolioRepository.findByIdAndClientId(id, securityUser.getId())
                 .orElseThrow(() -> new PortfolioNotFoundException(id));
         portfolio.setName(portfolioName);
         return portfolioRepository.save(portfolio);
@@ -33,15 +43,23 @@ public class PortfolioService {
         portfolioRepository.save(portfolio);
     }
 
-    public Portfolio getPortfolioById(Long id) {
-        return portfolioRepository.findById(id)
+    public Portfolio getPortfolioById(SecurityUser securityUser, Long id) {
+        return portfolioRepository.findByIdAndClientId(id, securityUser.getId())
                 .orElseThrow(() -> new PortfolioNotFoundException(id));
     }
 
-    public void deletePortfolio(Long id) {
-        portfolioRepository.findById(id)
-                        .orElseThrow(() -> new PortfolioNotFoundException(id));
+    @Transactional
+    public void deletePortfolio(SecurityUser securityUser, Long id) {
+        Portfolio portfolio = portfolioRepository
+                .findByIdAndClientId(id, securityUser.getId())
+                .orElseThrow(() -> new PortfolioNotFoundException(id));
+        if(!portfolio.getSecurities().isEmpty())
+            throw new PortfolioIsNotEmptyException(portfolio.getName());
         portfolioRepository.deleteById(id);
+    }
+
+    public Page<Portfolio> getAllPortfolios(SecurityUser user, Pageable pageable) {
+        return portfolioRepository.findAllByClientIdOrderByIdDesc(user.getId(), pageable);
     }
 
     public List<Portfolio> getAllPortfolios() {
